+++ src/modules/rest/rest.gateway.ts
@@ -8,7 +8,6 @@ import { Inject, Service } from '../../app/container';
 import { AuthenticationHandlerProvider } from '../authentication/provider/authentication-handler.provider';
 import { ContentAction } from '../content/interface/content-action.enum';
 import { IHttpGateway } from '../http/interface/http-gateway.interface';
-import { IODataResult } from '../odata/interface/odata-result.interface';
 import { SchemaService } from '../schema/service/schema.service';
 import { RestService } from './rest.service';
 
@@ -50,7 +49,7 @@ export class RestGateway implements IHttpGateway {
 
             reply.statusCode = 201;
 
-            return response;
+            return JSON.parse(JSON.stringify(response)); // TODO: need to convert the date into string before responding
           } catch (error) {
             reply.statusCode = 400;
 
@@ -83,14 +82,14 @@ export class RestGateway implements IHttpGateway {
           );
 
           if (record) {
-            return record;
+            return JSON.parse(JSON.stringify(record));
           }
 
           // Handle the 404 error
           reply.statusCode = 404;
 
           return {
-            error: 'Not found',
+            error: 'Not Found',
             statusCode: 404,
           };
         },
@@ -109,13 +108,37 @@ export class RestGateway implements IHttpGateway {
         async (
           req: FastifyRequest<{
             Body: object;
+            Params: Record<string, string>;
           }>,
-        ): Promise<IODataResult[]> => {
-          return this.service.update(
-            schema.database,
-            schema.reference,
-            req.body as any[],
-          );
+          reply: FastifyReply,
+        ): Promise<unknown> => {
+          try {
+            const record = await this.service.update(
+              schema.database,
+              schema.reference,
+              req.params,
+              req.body as any,
+            );
+
+            if (record) {
+              return JSON.parse(JSON.stringify(record));
+            }
+
+            // Handle the 404 error
+            reply.statusCode = 404;
+
+            return {
+              error: 'Not Found',
+              statusCode: 404,
+            };
+          } catch (error) {
+            reply.statusCode = 400;
+
+            return {
+              statusCode: 400,
+              error: 'Bad Request',
+            };
+          }
         },
       );
 
@@ -129,12 +152,36 @@ export class RestGateway implements IHttpGateway {
           ),
           preHandler,
         },
-        async (req: FastifyRequest): Promise<IODataResult> => {
-          return this.service.delete(
-            schema.database,
-            schema.reference,
-            req.query as Record<string, any>,
-          );
+        async (
+          req: FastifyRequest<{ Params: Record<string, string> }>,
+          reply: FastifyReply,
+        ): Promise<unknown> => {
+          try {
+            const record = await this.service.delete(
+              schema.database,
+              schema.reference,
+              req.params,
+            );
+
+            if (record) {
+              return JSON.parse(JSON.stringify(record));
+            }
+
+            // Handle the 404 error
+            reply.statusCode = 404;
+
+            return {
+              error: 'Not Found',
+              statusCode: 404,
+            };
+          } catch (error) {
+            reply.statusCode = 400;
+
+            return {
+              statusCode: 400,
+              error: 'Bad Request',
+            };
+          }
         },
       );
     }